StudyBud/requirements.txt
black==23.3.0
django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
flake8==6.0.0
pre-commit==3.2.2